[REFACTOR] Using BaseController; Parent Class of Controller Layer (REF-13)
Make sendResponseHttpByJson method at parent class and all the controller class inherit BaseController. Able to use sending Http Json with parent's method.
