+++ src/main/kotlin/com/entrip/controller/BaseController.kt
@@ -1,0 +1,23 @@
+package com.entrip.controller
+
+import com.entrip.domain.RestAPIMessages
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.http.MediaType
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.RestController
+import java.nio.charset.Charset
+
+@RestController
+class BaseController {
+    fun sendResponseHttpByJson(message: String, data: Any): ResponseEntity<RestAPIMessages> {
+        val restAPIMessages: RestAPIMessages = RestAPIMessages(
+            httpStatus = 200,
+            message = message,
+            data = data
+        )
+        val headers: HttpHeaders = HttpHeaders()
+        headers.contentType = MediaType("application", "json", Charset.forName("UTF-8"))
+        return ResponseEntity<RestAPIMessages>(restAPIMessages, headers, HttpStatus.OK)
+    }
+}
